+++ CHANGELOG.md
@@ -680,3 +680,7 @@ a
 
 ## 0.1.597
 * Added ZwapRadioButton widget
+
+## 0.1.598
+* Added color and activeColor distinction in ZwapRadioButton
+* Now ZwapInput.collapsed(...) will never show an empty dynamic label